Fix compiledb-gen-make when invoked from $PATH
The current code has a bug when compiledb-gen-make is run from $PATH, as
the dirname will resolve to ''. The correct solution is to use
"${BASH_SOURCE[0]}" which has the actual path of the script being
invoked, rather than "$0" which just has the invoked name. This also
fixes compiledb-gen-aosp which has the same bug.