make wrapper: Not so ellegant solution for bug #35
This hack is intended to make it possible/feasible to use
"make --dry-run --always-make" to speed up the compilation
commands extraction for autoconf/automake based build systems.
Without this ugly hack, particularly `--always-make` causes the
autoconf's configure script to be executed to every make target,
severely slowing down the process.
To work around this issue we use a mock/empty script to be used
in place of config.status and missing is they exist.

- Fixes #35

Signed-off-by: Nick Diego Yamane <nick@diegoyam.com>