Fix multiple_commands_oneline testcase for macOS.
* By using /bin/echo, the non-portable nature of echo -n is avoided.

The alternative is to redefine run_cmd to use:

    subprocess.check_output(['/bin/bash', '-c', cmd], **kwargs)

This will run commands with bourne shell rather than the default /bin/sh
specification. The multiple_commands_oneline testcase ran afoul of the
fact that "echo -n" is not portable. Notably, on macOS the builtin "echo"
in  sh does not accept the "-n" option.

See https://apple.stackexchange.com/questions/173836/why-echo-n-doesnt-work-in-this-script-on-mac-terminal

The Travis tests did not reveal this because on modern Debian/Ubuntu
systems /bin/sh symbolically links to the dash shell, which provides a
builtin "echo" that does accept "-n".