Avoid spurious error messages caused by substitution attempts.
The CommandProcessor gets confused by configure output lines of the form:

    checking whether make sets $(MAKE)... yes

It tries to make a command substitution by calling a comannd MAKE, which
most likely will fail and doesn't make sense at all.  While this doesn't
really hurt, it is confusing to see lines of the form

    /bin/sh: 1: MAKE: not found

on stderr when running compiledb.  We now simply skip such lines, avoiding
the above error message.  Admittedly this is a very special case, but a very
common one, so we better handle it.