Fix popen call to run make
When running make via utils.popen, there is an implicit 'shell=True'
which means you should pass the command as a string, not a list.

However, the make command was a list, which led to this command being
handled incorrectly.  The arguments to make were instead treated as
arguments to the shell, which meant that make did not behave as
intended.  In particular, '-n' would be missing, so make would actually
build everything even though it shouldn't.

Fix this by converting the list to a string for this popen call.  To
support this, added a new utility function cmd_join, which is
essentially shlex.join from the standard library (but a homebrew version
because shlex.join didn't appear until Python 3.8).